pyproject.toml
[project]
name = "document-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
fastapi = "^0.128.0"
sentence-transformers = "5.2.2"
faiss-cpu = "1.13.2"
pypdf = "6.6.2"
python-dotenv = "^1.2.1"
langchain-text-splitters = "^1.1.0"
langchain = "1.2.7"
langchain-community = "0.4.1"
langchain-openai = "1.1.7"
langchain-chroma = "^1.1.0"
uvicorn = "^0.40.0"
python-multipart = "^0.0.22"
langchain-huggingface = "^1.2.0"
accelerate = "^1.12.0"
pdfplumber = "^0.11.9"
sentencepiece = "^0.2.1"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
pillow = "^12.1.0"
streamlit = "^1.53.1"
openai-whisper = "^20250625"
langchain-ollama = "^1.0.1"
python-docx = "^1.2.0"
alembic = "^1.18.3"
psycopg2-binary = "^2.9.11"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
pydantic-settings = "^2.12.0"
email-validator = "^2.3.0"
passlib = {extras = ["argon2"], version = "^1.7.4"}
slowapi = "^0.1.9"
redis = "^7.2.0"
celery = "^5.6.2"
pytest = "^9.0.2"
docling = "^2.74.0"
langchain-groq = "^1.1.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
